requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
idna==2.9
importlib-metadata==1.6.0
lark-parser==0.8.5
more-itertools==8.2.0
murmurhash==1.0.2
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pkg-resources==0.0.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==3.0.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.9
zipp==3.1.0
